// src/app/api/model/applied-opportunity-content-dto-in.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AppliedOpportunityContentDtoIn { 
    appliedOpportunityId: number;
    contentCount?: number;
    contentCreationDate?: string;
    contentTypeId: number;
    description?: string;
    socialMediaLink?: string;
    submissionDate?: string;
    tags?: string;
    urls?: Array<string>;
}